requirements.txt
fastapi
uvicorn
langchain-groq
langchain
pydantic
pandas
python-dotenv
streamlit
plotly
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2